% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuperSurv-methods.R
\name{eval_times}
\alias{eval_times}
\alias{eval_times.default}
\alias{eval_times.SuperSurv}
\title{Access SuperSurv prediction evaluation times}
\usage{
eval_times(object, ...)

\method{eval_times}{default}(object, ...)

\method{eval_times}{SuperSurv}(object, ...)
}
\arguments{
\item{object}{A fitted object of class \code{"SuperSurv"}.}

\item{...}{Additional arguments ignored.}
}
\value{
A numeric vector of prediction evaluation times.
}
\description{
Returns the time grid used for the fitted \code{SuperSurv} predictions.
}
