% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explainability.R
\name{explain_shap}
\alias{explain_shap}
\title{Explain Predictions with Global SHAP (Kernel SHAP)}
\usage{
explain_shap(model, X_explain, X_background, nsim = 20, only_best = FALSE)
}
\arguments{
\item{model}{A fitted SuperSurv object OR a single wrapper output.}

\item{X_explain}{The dataset you want to explain (e.g., \code{X_test[1:10, ]}).}

\item{X_background}{The reference dataset for fastshap (e.g., \code{X_train[1:100, ]}).}

\item{nsim}{Number of simulations. Defaults to 20.}

\item{only_best}{Logical. If TRUE and model is SuperSurv, only explains the highest-weighted base learner.}
}
\value{
A data.frame of class \code{c("explain", "data.frame")} containing the calculated SHAP values. The columns correspond to the covariates in \code{X_explain}.
}
\description{
Explain Predictions with Global SHAP (Kernel SHAP)
}
