% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmst.R
\name{get_rmst}
\alias{get_rmst}
\title{Calculate Restricted Mean Survival Time (RMST)}
\usage{
get_rmst(surv_matrix, times, tau)
}
\arguments{
\item{surv_matrix}{Matrix of survival probabilities (rows: patients, cols: time points)}

\item{times}{Vector of time points corresponding to the columns}

\item{tau}{The restriction time horizon}
}
\value{
A vector of RMST values for each patient
}
\description{
Calculate Restricted Mean Survival Time (RMST)
}
\keyword{internal}
