% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_wrappers.R
\name{list_wrappers}
\alias{list_wrappers}
\title{List Available Wrappers and Screeners in SuperSurv}
\usage{
list_wrappers(what = "both")
}
\arguments{
\item{what}{Character string. If \code{"both"} (default), lists both prediction
and screening functions. If \code{"surv"}, lists only prediction models.
If \code{"screen"}, lists only screening algorithms. Otherwise, lists all exports.}
}
\value{
An invisible character vector containing the requested function names.
}
\description{
This function prints all built-in prediction algorithms and feature screening
algorithms available in the \code{SuperSurv} package.
}
