% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metabric}
\alias{metabric}
\title{METABRIC Breast Cancer Dataset}
\format{
A data frame with the clinical and genomic variables:
\describe{
\item{duration}{Survival time.}
\item{event}{Event indicator (1 = event, 0 = censored).}
\item{x0}{Feature x0.}
\item{x1}{Feature x1.}
\item{x2}{Feature x2.}
\item{x3}{Feature x3.}
\item{x4}{Feature x4.}
\item{x5}{Feature x5.}
\item{x6}{Feature x6.}
\item{x7}{Feature x7.}
\item{x8}{Feature x8.}
}
}
\usage{
metabric
}
\description{
A subset of the Molecular Taxonomy of Breast Cancer International Consortium
(METABRIC) dataset to demonstrate the SuperSurv package.
}
\keyword{datasets}
