% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explainability.R
\name{plot_dependence}
\alias{plot_dependence}
\title{Plot SHAP Dependence for SuperSurv}
\usage{
plot_dependence(shap_values, data, feature_name, title = NULL)
}
\arguments{
\item{shap_values}{The output from \code{explain_kernel()}.}

\item{data}{The original covariate data used for the explanation (X_explain)}

\item{feature_name}{String name of the column to plot}

\item{title}{Optional custom title.}
}
\value{
A \code{ggplot} object visualizing the SHAP values.
}
\description{
Plot SHAP Dependence for SuperSurv
}
\examples{
if (FALSE) {
  data("metabric", package = "SuperSurv")
  dat <- metabric[1:80, ]
  x_cols <- grep("^x", names(dat))[1:5]
  X <- dat[, x_cols, drop = FALSE]
  new.times <- seq(20, 120, by = 20)

  fit <- SuperSurv(
    time = dat$duration,
    event = dat$event,
    X = X,
    newdata = X,
    new.times = new.times,
    event.library = c("surv.coxph", "surv.ridge"),
    cens.library = c("surv.coxph"),
    control = list(saveFitLibrary = TRUE)
  )

  shap_values <- explain_kernel(
    model = fit,
    X_explain = X[1:20, , drop = FALSE],
    X_background = X[21:50, , drop = FALSE],
    nsim = 5, eval_time = 100
  )

  plot_dependence(
    shap_values = shap_values,
    data = X[1:20, , drop = FALSE],
    feature_name = colnames(X)[1]
  )
}
}
