% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explainability.R
\name{plot_global_importance}
\alias{plot_global_importance}
\title{Plot Global Feature Importance for SuperSurv}
\usage{
plot_global_importance(
  shap_values,
  title = "SuperSurv: Ensemble Feature Importance",
  top_n = 10
)
}
\arguments{
\item{shap_values}{The output from \code{explain_kernel()}.}

\item{title}{Plot title.}

\item{top_n}{Number of features to show (default 10)}
}
\value{
A \code{ggplot} object visualizing the SHAP values.
}
\description{
Plot Global Feature Importance for SuperSurv
}
\examples{
if (FALSE) {
  data("metabric", package = "SuperSurv")
  dat <- metabric[1:80, ]
  x_cols <- grep("^x", names(dat))[1:5]
  X <- dat[, x_cols, drop = FALSE]
  new.times <- seq(20, 120, by = 20)

  fit <- SuperSurv(
    time = dat$duration,
    event = dat$event,
    X = X,
    newdata = X,
    new.times = new.times,
    event.library = c("surv.coxph", "surv.ridge"),
    cens.library = c("surv.coxph"),
    control = list(saveFitLibrary = TRUE)
  )

  shap_values <- explain_kernel(
    model = fit,
    X_explain = X[1:10, , drop = FALSE],
    X_background = X[11:40, , drop = FALSE],
    nsim = 5, eval_time = 100
  )

  plot_global_importance(shap_values, top_n = 5)
}
}
