% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_performance_comparison.R
\name{plot_performance_comparison}
\alias{plot_performance_comparison}
\title{Benchmark Ensemble vs. Baseline}
\usage{
plot_performance_comparison(explainers)
}
\arguments{
\item{explainers}{A list of survex explainers (e.g., list(Ensemble=e1, Cox=e2))}
}
\description{
Benchmark Ensemble vs. Baseline
}
