% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_survival_heatmap.R
\name{plot_survival_heatmap}
\alias{plot_survival_heatmap}
\title{Survival Probability Heatmap}
\usage{
plot_survival_heatmap(object, newdata, times)
}
\arguments{
\item{object}{A fitted SuperSurv object}

\item{newdata}{Test covariates (e.g., \code{X_te[1:50, ]})}

\item{times}{The time grid to visualize}
}
\value{
A \code{ggplot} object visualizing the SHAP values.
}
\description{
Survival Probability Heatmap
}
