% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{predict.surv.bart}
\alias{predict.surv.bart}
\title{Prediction function for BART}
\usage{
\method{predict}{surv.bart}(object, newdata, new.times, ...)
}
\arguments{
\item{object}{Fitted \code{surv.bart} object.}

\item{newdata}{New covariate data.frame to predict on.}

\item{new.times}{Times to predict.}

\item{...}{Additional ignored arguments.}
}
\value{
A numeric matrix of predicted survival probabilities, where rows correspond
to the observations in \code{newdata} and columns correspond to the evaluation
times in \code{new.times}.
}
\description{
Obtains predicted survivals from a fitted \code{surv.bart} object.
Manually expands the time grid to bypass the 10-column expectation.
}
