% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{predict.surv.coxph}
\alias{predict.surv.coxph}
\title{Prediction function for Cox regression wrapper}
\usage{
\method{predict}{surv.coxph}(object, newdata, new.times, ...)
}
\arguments{
\item{object}{Fitted \code{surv.coxph} object.}

\item{newdata}{New covariate data.frame for which to obtain predictions.}

\item{new.times}{Times at which to obtain predicted survivals.}

\item{...}{Additional ignored arguments.}
}
\value{
A numeric matrix of predicted survival probabilities, where rows correspond
to the observations in \code{newdata} and columns correspond to the evaluation
times in \code{new.times}.
}
\description{
Obtains predicted survivals from a fitted \code{surv.coxph} object.
}
\examples{
data("metabric", package = "SuperSurv")
dat <- metabric[1:30, ]
x_cols <- grep("^x", names(dat))[1:3]
X <- dat[, x_cols, drop = FALSE]
newX <- X[1:5, , drop = FALSE]
times <- seq(50, 150, by = 50)

fit <- surv.coxph(
  time = dat$duration,
  event = dat$event,
  X = X,
  newdata = newX,
  new.times = times,
  obsWeights = rep(1, nrow(dat)),
  id = NULL
)

pred <- predict(fit$fit, newdata = newX, new.times = times)
dim(pred)
}
