% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{predict.surv.km}
\alias{predict.surv.km}
\title{Predict Method for Kaplan-Meier Wrapper}
\usage{
\method{predict}{surv.km}(object, newdata, new.times, ...)
}
\arguments{
\item{object}{A fitted object of class \code{surv.km}.}

\item{newdata}{New covariate data.frame for which to obtain predictions (Ignored).}

\item{new.times}{Numeric vector of times at which to predict survival.}

\item{...}{Additional ignored arguments.}
}
\value{
A numeric matrix of predicted survival probabilities, where rows correspond
to the observations in \code{newdata} and columns correspond to the evaluation times.
}
\description{
Obtains predicted survivals from a fitted \code{surv.km} object.
}
\examples{
data("metabric", package = "SuperSurv")
dat <- metabric[1:30, ]
x_cols <- grep("^x", names(dat))[1:3]
X <- dat[, x_cols, drop = FALSE]
newX <- X[1:5, , drop = FALSE]
times <- seq(50, 150, by = 50)

fit <- surv.km(
  time = dat$duration,
  event = dat$event,
  X = X,
  newdata = newX,
  new.times = times,
  obsWeights = rep(1, nrow(dat)),
  id = NULL
)

pred <- predict(fit$fit, newdata = newX, new.times = times)
dim(pred)
}
