% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screener.R
\name{screen.all}
\alias{screen.all}
\title{Keep All Variables Screener}
\usage{
screen.all(X, ...)
}
\arguments{
\item{X}{Training covariate data.frame.}

\item{...}{Additional ignored arguments.}
}
\value{
A logical vector of the same length as the number of columns in \code{X},
indicating which variables passed the screening algorithm (\code{TRUE} to keep,
\code{FALSE} to drop).
}
\description{
Keep All Variables Screener
}
\examples{
data("metabric", package = "SuperSurv")
dat <- metabric[1:20, ]
x_cols <- grep("^x", names(dat))[1:5]
X <- dat[, x_cols, drop = FALSE]

screen.all(X)
}
