% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screeners.R
\name{screen.marg}
\alias{screen.marg}
\title{Marginal Cox Regression Screening}
\usage{
screen.marg(time, event, X, obsWeights, minscreen = 2, min.p = 0.1, ...)
}
\arguments{
\item{time}{Observed follow-up time.}

\item{event}{Observed event indicator.}

\item{X}{Training covariate data.frame.}

\item{obsWeights}{Observation weights.}

\item{minscreen}{Minimum number of covariates to return. Defaults to 2.}

\item{min.p}{Threshold p-value. Defaults to 0.1.}

\item{...}{Additional ignored arguments.}
}
\value{
A logical vector of the same length as the number of columns in \code{X},
indicating which variables passed the screening algorithm (\code{TRUE} to keep,
\code{FALSE} to drop).
}
\description{
Marginal Cox Regression Screening
}
