% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuperSurv.CV.control.R
\name{SuperSurv.CV.control}
\alias{SuperSurv.CV.control}
\title{Control parameters for Cross-Validation in SuperSurv}
\usage{
SuperSurv.CV.control(
  V = 10L,
  stratifyCV = TRUE,
  shuffle = TRUE,
  validRows = NULL
)
}
\arguments{
\item{V}{Number of folds. Default is 10.}

\item{stratifyCV}{Logical. If TRUE, ensures event rates are balanced across folds.}

\item{shuffle}{Logical. If TRUE, shuffles rows before splitting.}

\item{validRows}{Optional custom list of indices for folds.}
}
\value{
A list of CV parameters.
}
\description{
Control parameters for Cross-Validation in SuperSurv
}
