% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{surv.exponential}
\alias{surv.exponential}
\title{Parametric Survival Prediction Wrapper (Exponential)}
\usage{
surv.exponential(time, event, X, newdata, new.times, obsWeights, id, ...)
}
\arguments{
\item{time}{Observed follow-up time.}

\item{event}{Observed event indicator.}

\item{X}{Training covariate data.frame.}

\item{newdata}{Test covariate data.frame to use for prediction.}

\item{new.times}{Times at which to obtain the predicted survivals.}

\item{obsWeights}{Observation weights.}

\item{id}{Cluster identification variable.}

\item{...}{Additional ignored arguments.}
}
\value{
A list containing the fitted model and predictions.
}
\description{
Parametric Survival Prediction Wrapper (Exponential)
}
\examples{
data("metabric", package = "SuperSurv")
dat <- metabric[1:30, ]
x_cols <- grep("^x", names(dat))[1:3]
X <- dat[, x_cols, drop = FALSE]
newX <- X[1:5, , drop = FALSE]
times <- seq(50, 150, by = 50)

fit <- surv.exponential(
  time = dat$duration,
  event = dat$event,
  X = X,
  newdata = newX,
  new.times = times,
  obsWeights = rep(1, nrow(dat)),
  id = NULL
)

dim(fit$pred)
}
