% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{surv.loglogistic}
\alias{surv.loglogistic}
\title{Parametric Survival Prediction Wrapper (Log-Logistic)}
\usage{
surv.loglogistic(time, event, X, newdata, new.times, obsWeights, id, ...)
}
\arguments{
\item{time}{Observed follow-up time.}

\item{event}{Observed event indicator.}

\item{X}{Training covariate data.frame.}

\item{newdata}{Test covariate data.frame to use for prediction.}

\item{new.times}{Times at which to obtain the predicted survivals.}

\item{obsWeights}{Observation weights.}

\item{id}{Cluster identification variable.}

\item{...}{Additional ignored arguments.}
}
\value{
A list containing the fitted model and predictions.
}
\description{
Parametric Survival Prediction Wrapper (Log-Logistic)
}
