% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{eval_cindex}
\alias{eval_cindex}
\title{Calculate Concordance Index (Harrell's or Uno's)}
\usage{
eval_cindex(time, event, S_mat, times, eval_time, method = "uno")
}
\arguments{
\item{time}{Numeric vector of observed follow-up times.}

\item{event}{Numeric vector of event indicators (1 = event, 0 = censored).}

\item{S_mat}{A numeric matrix of predicted survival probabilities.}

\item{times}{Numeric vector of evaluation times matching the columns of \code{S_mat}.}

\item{eval_time}{Numeric. The specific time point at which to extract predictions.}

\item{method}{Character. Either "harrell" or "uno". Defaults to "uno".}
}
\value{
A numeric value representing the chosen C-index.
}
\description{
Calculate Concordance Index (Harrell's or Uno's)
}
