% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuperSurv-methods.R
\name{learner_names}
\alias{learner_names}
\alias{learner_names.SuperSurv}
\title{Access SuperSurv learner names}
\usage{
learner_names(object, ...)

\method{learner_names}{SuperSurv}(object, type = c("event", "censoring", "both"), ...)
}
\arguments{
\item{object}{A fitted object of class \code{"SuperSurv"}.}

\item{...}{Additional arguments ignored.}

\item{type}{Character string specifying whether to return event learner
names, censoring learner names, or both.}
}
\value{
For \code{type = "event"} or \code{type = "censoring"}, a character
vector of learner names. For \code{type = "both"}, a list with elements
\code{event} and \code{censoring}.
}
\description{
Returns the fitted learner names from a \code{SuperSurv} object.
}
