% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_beeswarm.R
\name{plot_beeswarm}
\alias{plot_beeswarm}
\title{Beeswarm Summary Plot for SuperSurv SHAP}
\usage{
plot_beeswarm(shap_values, data, top_n = 10)
}
\arguments{
\item{shap_values}{The output from explain_shap.SuperSurv}

\item{data}{The covariate data used (X_explain)}

\item{top_n}{Number of features to display}
}
\value{
A \code{ggplot} object visualizing the SHAP values.
}
\description{
Beeswarm Summary Plot for SuperSurv SHAP
}
