% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dependence.R
\name{plot_dependence}
\alias{plot_dependence}
\title{Plot SHAP Dependence for SuperSurv}
\usage{
plot_dependence(shap_values, data, feature_name, title = NULL)
}
\arguments{
\item{shap_values}{The output from explain_shap.SuperSurv}

\item{data}{The original covariate data used for the explanation (X_explain)}

\item{feature_name}{String name of the column to plot}

\item{title}{Optional custom title.}
}
\value{
A \code{ggplot} object visualizing the SHAP values.
}
\description{
Plot SHAP Dependence for SuperSurv
}
