% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_global_importance.R
\name{plot_global_importance}
\alias{plot_global_importance}
\title{Plot Global Feature Importance for SuperSurv}
\usage{
plot_global_importance(
  shap_values,
  title = "SuperSurv: Ensemble Feature Importance",
  top_n = 10
)
}
\arguments{
\item{shap_values}{The output from explain_shap.SuperSurv}

\item{title}{Plot title.}

\item{top_n}{Number of features to show (default 10)}
}
\value{
A \code{ggplot} object visualizing the SHAP values.
}
\description{
Plot Global Feature Importance for SuperSurv
}
