% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmst.R
\name{plot_marginal_rmst_curve}
\alias{plot_marginal_rmst_curve}
\title{Plot Causal RMST Difference Over Time}
\usage{
plot_marginal_rmst_curve(fit, data, trt_col, times, tau_seq)
}
\arguments{
\item{fit}{A fitted \code{SuperSurv} ensemble object.}

\item{data}{A \code{data.frame} containing the patient covariates and the treatment assignment.}

\item{trt_col}{Character string. The exact name of the binary treatment indicator column in \code{data}.}

\item{times}{Numeric vector of time points matching the prediction grid.}

\item{tau_seq}{Numeric vector. A sequence of truncation times (\code{tau}) to evaluate and plot.}
}
\value{
A \code{ggplot} object visualizing the causal RMST difference curve.
}
\description{
Generates a curve showing how the causal Restricted Mean Survival Time (RMST) difference
between treatment groups evolves across a sequence of different truncation times.
}
\examples{
data("metabric", package = "SuperSurv")
dat <- metabric[1:80, ]
x_cols <- grep("^x", names(dat))[1:5]
X <- dat[, x_cols, drop = FALSE]
new.times <- seq(20, 120, by = 20)

fit <- SuperSurv(
  time = dat$duration,
  event = dat$event,
  X = X,
  newdata = X,
  new.times = new.times,
  event.library = c("surv.coxph", "surv.glmnet"),
  cens.library = c("surv.coxph"),
  control = list(saveFitLibrary = TRUE)
)

tau_grid <- seq(40, 120, by = 20)
plot_marginal_rmst_curve(
  fit = fit,
  data = dat,
  trt_col = "x4",
  times = new.times,
  tau_seq = tau_grid
)
}
