% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_patient_waterfall.R
\name{plot_patient_waterfall}
\alias{plot_patient_waterfall}
\title{Waterfall Plot for an Individual Patient}
\usage{
plot_patient_waterfall(shap_values, patient_index = 1, top_n = 10)
}
\arguments{
\item{shap_values}{The output from explain_shap.*}

\item{patient_index}{The row index of the patient to explain}

\item{top_n}{Number of features to show (default 10)}
}
\value{
A \code{ggplot} object visualizing the SHAP values.
}
\description{
Waterfall Plot for an Individual Patient
}
