% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_predict}
\alias{plot_predict}
\title{Plot Predicted Survival Curves}
\usage{
plot_predict(preds, eval_times, patient_idx = 1)
}
\arguments{
\item{preds}{A list containing predictions, specifically the object returned
by the \code{predict.SuperSurv} function.}

\item{eval_times}{Numeric vector of times at which predictions were evaluated.}

\item{patient_idx}{Integer vector. The row indices of the patients in the test
dataset whose survival curves you want to plot. Defaults to 1.}
}
\value{
A \code{ggplot} object showing the survival curves.
}
\description{
Generates a ggplot2 step-function plot of the predicted survival probabilities
over time for one or more specific patients.
}
