% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv.aorsf.R
\name{predict.surv.aorsf}
\alias{predict.surv.aorsf}
\title{Prediction function for AORSF}
\usage{
\method{predict}{surv.aorsf}(object, newX, new.times, ...)
}
\arguments{
\item{object}{Fitted \code{surv.aorsf} object.}

\item{newX}{New covariate data.frame.}

\item{new.times}{Times at which to obtain predicted survivals.}

\item{...}{Additional ignored arguments.}
}
\value{
A numeric matrix of predicted survival probabilities, where rows correspond
to the observations in \code{newX} and columns correspond to the evaluation
times in \code{new.times}.
}
\description{
Obtains predicted survivals from a fitted \code{surv.aorsf} object.
Uses the native \code{aorsf} prediction engine to calculate survival directly at requested times.
}
\keyword{internal}
