% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv.coxph.R
\name{predict.surv.coxph}
\alias{predict.surv.coxph}
\title{Prediction function for Cox regression wrapper}
\usage{
\method{predict}{surv.coxph}(object, newX, new.times, ...)
}
\arguments{
\item{object}{Fitted \code{surv.coxph} object.}

\item{newX}{New covariate data.frame for which to obtain predictions.}

\item{new.times}{Times at which to obtain predicted survivals.}

\item{...}{Additional ignored arguments.}
}
\value{
A numeric matrix of predicted survival probabilities, where rows correspond
to the observations in \code{newX} and columns correspond to the evaluation
times in \code{new.times}.
}
\description{
Obtains predicted survivals from a fitted \code{surv.coxph} object.
}
