% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{predict.surv.gam}
\alias{predict.surv.gam}
\title{Prediction function for GAM wrapper}
\usage{
\method{predict}{surv.gam}(object, newdata, new.times, ...)
}
\arguments{
\item{object}{Fitted \code{surv.gam} object.}

\item{newdata}{New covariate data.frame.}

\item{new.times}{Times at which to obtain predicted survivals.}

\item{...}{Additional ignored arguments.}
}
\value{
A numeric matrix of predicted survival probabilities, where rows correspond
to the observations in \code{newdata} and columns correspond to the evaluation
times in \code{new.times}.
}
\description{
Obtains predicted survivals from a fitted \code{surv.gam} object.
}
\examples{
if (requireNamespace("mgcv", quietly = TRUE)) {
  data("metabric", package = "SuperSurv")
  dat <- metabric[1:30, ]
  x_cols <- grep("^x", names(dat))[1:3]
  X <- dat[, x_cols, drop = FALSE]
  newX <- X[1:5, , drop = FALSE]
  times <- seq(50, 150, by = 50)

  fit <- surv.gam(
    time = dat$duration,
    event = dat$event,
    X = X,
    newdata = newX,
    new.times = times,
    obsWeights = rep(1, nrow(dat)),
    id = NULL,
    cts.num = 5
  )

  pred <- predict(fit$fit, newdata = newX, new.times = times)
  dim(pred)
}
}
