% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv.gam.R
\name{predict.surv.gam}
\alias{predict.surv.gam}
\title{Prediction function for GAM wrapper}
\usage{
\method{predict}{surv.gam}(object, newX, new.times, ...)
}
\arguments{
\item{object}{Fitted \code{surv.gam} object.}

\item{newX}{New covariate data.frame.}

\item{new.times}{Times at which to obtain predicted survivals.}

\item{...}{Additional ignored arguments.}
}
\value{
A numeric matrix of predicted survival probabilities, where rows correspond
to the observations in \code{newX} and columns correspond to the evaluation
times in \code{new.times}.
}
\description{
Obtains predicted survivals from a fitted \code{surv.gam} object.
}
