% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{predict.surv.km}
\alias{predict.surv.km}
\title{Predict Method for Kaplan-Meier Wrapper}
\usage{
\method{predict}{surv.km}(object, newdata, new.times, ...)
}
\arguments{
\item{object}{A fitted object of class \code{surv.km}.}

\item{newdata}{New covariate data.frame for which to obtain predictions (Ignored).}

\item{new.times}{Numeric vector of times at which to predict survival.}

\item{...}{Additional ignored arguments.}
}
\value{
A numeric matrix of predicted survival probabilities, where rows correspond
to the observations in \code{newdata} and columns correspond to the evaluation times.
}
\description{
Obtains predicted survivals from a fitted \code{surv.km} object.
}
\keyword{internal}
