% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv.parametric.R
\name{predict.surv.parametric}
\alias{predict.surv.parametric}
\title{Prediction function for Universal Parametric Wrapper}
\usage{
\method{predict}{surv.parametric}(object, newX, new.times, ...)
}
\arguments{
\item{object}{Fitted \code{surv.parametric} object.}

\item{newX}{New covariate data.frame.}

\item{new.times}{Times at which to obtain predicted survivals.}

\item{...}{Additional ignored arguments.}
}
\value{
A numeric matrix of predicted survival probabilities, where rows correspond
to the observations in \code{newX} and columns correspond to the evaluation
times in \code{new.times}.
}
\description{
Obtains predicted survivals from a fitted \code{surv.parametric} object
using exact closed-form equations.
}
