% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{predict.surv.ranger}
\alias{predict.surv.ranger}
\title{Prediction function for Ranger wrapper}
\usage{
\method{predict}{surv.ranger}(object, newdata, new.times, ...)
}
\arguments{
\item{object}{Fitted \code{surv.ranger} object.}

\item{newdata}{New covariate data.frame.}

\item{new.times}{Times at which to obtain predicted survivals.}

\item{...}{Additional ignored arguments.}
}
\value{
A numeric matrix of predicted survival probabilities, where rows correspond
to the observations in \code{newdata} and columns correspond to the evaluation
times in \code{new.times}.
}
\description{
Obtains predicted survivals from a fitted \code{surv.ranger} object.
}
