% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screeners.R
\name{screen.all}
\alias{screen.all}
\title{Keep All Variables Screener}
\usage{
screen.all(X, ...)
}
\arguments{
\item{X}{Training covariate data.frame.}

\item{...}{Additional ignored arguments.}
}
\value{
A logical vector of the same length as the number of columns in \code{X},
indicating which variables passed the screening algorithm (\code{TRUE} to keep,
\code{FALSE} to drop).
}
\description{
Keep All Variables Screener
}
