% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuperSurv.control.R
\name{SuperSurv.control}
\alias{SuperSurv.control}
\title{Control parameters for the SuperSurv Ensemble}
\usage{
SuperSurv.control(
  max.SL.iter = 20,
  event.t.grid = NULL,
  cens.t.grid = NULL,
  saveFitLibrary = TRUE,
  initWeightAlg = "surv.coxph",
  initWeight = "censoring",
  ...
)
}
\arguments{
\item{max.SL.iter}{Maximum iterations for the iterative weighting algorithm. Default 20.}

\item{event.t.grid}{Optional time grid for event risk calculation.}

\item{cens.t.grid}{Optional time grid for censoring risk calculation.}

\item{saveFitLibrary}{Logical. If TRUE (default), saves models for future predictions.}

\item{initWeightAlg}{The learner used for the very first step of IPCW.}

\item{initWeight}{Whether to start by fitting "censoring" or "event" weights.}

\item{...}{Additional ignored arguments.}
}
\value{
A list of control parameters.
}
\description{
Control parameters for the SuperSurv Ensemble
}
