% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuperSurv.R
\name{SuperSurv}
\alias{SuperSurv}
\title{Super Learner for conditional survival functions}
\usage{
SuperSurv(
  time,
  event,
  X,
  newX = NULL,
  new.times,
  event.SL.library,
  cens.SL.library,
  id = NULL,
  verbose = FALSE,
  control = list(),
  cvControl = list(),
  obsWeights = NULL,
  metalearner = "least_squares",
  selection = "ensemble",
  nFolds = 10,
  parallel = FALSE
)
}
\arguments{
\item{time}{Observed follow-up time.}

\item{event}{Observed event indicator.}

\item{X}{Training covariate data.frame.}

\item{newX}{Test covariate data.frame for prediction (defaults to X).}

\item{new.times}{Times at which to obtain predicted survivals.}

\item{event.SL.library}{Character vector of prediction algorithms for the event.}

\item{cens.SL.library}{Character vector of prediction algorithms for censoring.}

\item{id}{Cluster identification variable.}

\item{verbose}{Logical. If TRUE, prints progress messages.}

\item{control}{List of control parameters for the Super Learner.}

\item{cvControl}{List of control parameters for cross-validation.}

\item{obsWeights}{Observation weights.}

\item{metalearner}{Character string specifying the optimizer (e.g., "least_squares").}

\item{selection}{Character. Specifies how the meta-learner combines the base models.
Use \code{"ensemble"} (default) to calculate a weighted average (convex combination)
of the base learners. Use \code{"best"} to act as a Discrete Super Learner, which
assigns a weight of 1.0 to the single model with the lowest cross-validated risk.}

\item{nFolds}{Number of cross-validation folds (default: 10).}

\item{parallel}{Logical. If TRUE, uses future.apply for parallel execution.}
}
\value{
A list of class \code{SuperSurv} containing:
\itemize{
\item \code{call}: The matched function call.
\item \code{event.SL.predict}: Matrix of in-sample cross-validated survival predictions.
\item \code{cens.SL.predict}: Matrix of in-sample cross-validated censoring predictions.
\item \code{event.coef}: Numeric vector of optimized ensemble weights for the event.
\item \code{cens.coef}: Numeric vector of optimized ensemble weights for censoring.
\item \code{event.library.predict}: 3D array of cross-validated predictions from individual event learners.
\item \code{event.libraryNames}: Data frame detailing the algorithms and screeners used.
\item \code{event.fitLibrary}: List of the fitted base learner models (if \code{saveFitLibrary = TRUE}).
\item \code{times}: The time grid used for evaluation.
}
}
\description{
Orchestrates the cross-validation, metalearner optimization, and prediction
for an ensemble of survival base learners.
}
