% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuperSurv-methods.R
\name{training_variables}
\alias{training_variables}
\alias{training_variables.SuperSurv}
\title{Access SuperSurv training variable names}
\usage{
training_variables(object, ...)

\method{training_variables}{SuperSurv}(object, ...)
}
\arguments{
\item{object}{A fitted object of class \code{"SuperSurv"}.}

\item{...}{Additional arguments ignored.}
}
\value{
A character vector of training variable names.
}
\description{
Returns the covariate names used to fit a \code{SuperSurv} object.
}
